Extracted BytesGenerator protocol
